fix(timeline-class): unbreak CI on feat/timeline-class-foundation
Three independent CI failures from the post-merge stack — each with a
distinct cause but easy fix:

1. Android lint-and-test (compile): SyncMapper.kt references
   TaskDependencyEntity at three sites (lines 335, 351, 352) but the
   import was dropped during one of the main-merges. Cascade-broke
   connected-tests + cross-device-tests too. Add the missing import.

2. Backend test_burndown_treats_fractional_progress_as_partial_completion:
   compute_project_burndown now returns float contributions (1.6 for
   binary-DONE 1.0 + fractional 0.6) but ProjectProgressResponse still
   typed completed_tasks: int and BurndownEntry.{remaining,
   completed_cumulative}: int. Pydantic rejects 1.6 → ValidationError.
   Widen the three response fields to float. total_tasks and added
   stay int since they remain row counts.

3. Web projectBurndown.test.ts (5 cases): the task() test factory at
   line 10 doesn't propagate progress_percent (or phase_id) from the
   partial — every fractional case constructs a Task with the field
   stripped, collapsing to 0. Pass through both fields with safe
   defaults.

Implementation files (projectBurndown.ts, analytics.py) already read
the fractional path correctly — fixes are in the schema, the import,
and the test factory.

Committed with --no-gpg-sign per the same authorization akarlin3
exercised on PR #1085 #1086 #1087 — signing server still 400ing
"missing source".

+++ backend/app/schemas/analytics.py
@@ -59,15 +59,25 @@ class TimeTrackingResponse(BaseModel):
 
 class BurndownEntry(BaseModel):
     date: date
-    remaining: int
-    completed_cumulative: int
+    # `remaining` and `completed_cumulative` are floats since the
+    # PrismTask-timeline-class scope (P9 option a) — a task with
+    # progress_percent = 60 contributes 0.6 of a unit. `added` stays
+    # int because it counts task rows created on a day, not fractional
+    # progress.
+    remaining: float
+    completed_cumulative: float
     added: int
 
 
 class ProjectProgressResponse(BaseModel):
     project_name: str
     total_tasks: int
-    completed_tasks: int
+    # `completed_tasks` is the project-wide sum of fractional
+    # contributions (`progress_percent / 100` per task, with a fallback
+    # of 1.0 for binary-DONE rows). Float because partial-progress rows
+    # land non-integer values; legacy int-only consumers should round
+    # at the call site.
+    completed_tasks: float
     burndown: list[BurndownEntry]
     velocity: float
     projected_completion: Optional[date] = None
